servers/image_fetch_mcp/requirements.txt
mcp
fastmcp
requests
python-dotenv
